added other params to the network call, also added italian language to the query

+++ app/src/main/java/com/example/apiauthtest/Current.kt
@@ -33,7 +33,7 @@ data class Current(
     @SerializedName("pressure_mb")
     val pressureMb: Int?,
     @SerializedName("temp_c")
-    val tempC: Int?,
+    val tempC: Double?,
     @SerializedName("temp_f")
     val tempF: Double?,
     @SerializedName("uv")
